Add separate applications config for tests.
Spring-boot will load the application config automatically if included in the test/resources folder.
These are then loaded automatically into the ADB class. Spring-boot can not however do this into a static class. So the structure is changed so the ADB is a service that is injected into every DAO.

The yaml structure is pure because I think it's nice. This can be changed to a .properties file if the team disagrees.